added authentication functionalities.

+++ app/Http/Controllers/ideaController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Idea;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 
 class IdeaController extends Controller
@@ -17,19 +18,24 @@ class IdeaController extends Controller
 
     public function store()
     {
-        request()->validate([
+        $user_id = Auth::id();
+        $validated = request()->validate([
             'content' => 'required|min:5',
         ]);
-
+    
         Idea::create([
             'content' => request()->get('content', ''),
+            'user_id' => $user_id,
         ]);
-
+    
         return redirect()->route('dashboard')->with('success', 'Your post has been added successfully.');
     }
-
+    
     public function destroy($id)
     {
+        if(Auth::user()->id !== $id){
+            abort(404);
+        }
         $idea = Idea::findOrFail($id);
         $idea->delete();
 
@@ -38,6 +44,9 @@ class IdeaController extends Controller
 
     public function edit($id)
     {
+        if(Auth::user()->id !== $id){
+            abort(404);
+        }
         $idea = Idea::findOrFail($id);
         $editing = true;
 
@@ -46,6 +55,9 @@ class IdeaController extends Controller
 
     public function update($id)
     {
+        if(Auth::user()->id !== $id){
+            abort(404);
+        }
         request()->validate([
             'content' => 'required|min:5',
         ]);
@@ -56,4 +68,7 @@ class IdeaController extends Controller
 
         return redirect()->route('idea.show', $idea->id)->with('success', 'Idea updated successfully.');
     }
+
+   
+
 }
